Separate unit and performance testing
