+++ test/unit/acts_as_url/adapter/active_record.rb
