+++ test/unit/acts_as_url/adapter/data_mapper.rb
