+++ test/unit/acts_as_url/adapter/mongoid.rb
