+++ test/unit/acts_as_url_configuration_test.rb
