+++ test/unit/acts_as_url_integration_test.rb
