+++ test/unit/string_extensions_test.rb
