+++ test/unit/unicode_point_suite/codepoint_test_helper.rb
